+++ src/main/java/com/lowdragmc/mbd2/api/capability/recipe/RecipeCapability.java
@@ -6,8 +6,6 @@ import com.lowdragmc.lowdraglib.jei.IngredientIO;
 import com.lowdragmc.mbd2.api.recipe.content.Content;
 import com.lowdragmc.mbd2.api.recipe.content.ContentModifier;
 import com.lowdragmc.mbd2.api.recipe.content.IContentSerializer;
-import io.netty.buffer.Unpooled;
-import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.Component;
 
 import java.util.List;
@@ -32,9 +30,7 @@ public abstract class RecipeCapability<T> {
      * it's not recommended to use this method directly, use {@link #copyContent(Object)} instead.
      */
     public final T deepCopyContent(Object content) {
-        var buf = new FriendlyByteBuf(Unpooled.buffer());
-        serializer.toNetwork(buf, (T) content);
-        return serializer.fromNetwork(buf);
+        return serializer.deepCopyInner((T) content);
     }
 
     public final T deepCopyContent(Object content, ContentModifier modifier) {