+++ src/main/java/com/lowdragmc/mbd2/api/recipe/content/SerializerDouble.java
@@ -68,4 +68,8 @@ public class SerializerDouble implements IContentSerializer<Double> {
         return content;
     }
 
+    @Override
+    public Double deepCopyInner(Double content) {
+        return content;
+    }
 }