+++ src/main/java/com/lowdragmc/mbd2/api/recipe/content/SerializerIngredient.java
@@ -58,4 +58,9 @@ public class SerializerIngredient implements IContentSerializer<Ingredient> {
     public Ingredient copyWithModifier(Ingredient content, ContentModifier modifier) {
         return content instanceof SizedIngredient sizedIngredient ? SizedIngredient.create(sizedIngredient.getInner(), modifier.apply(sizedIngredient.getAmount()).intValue()) : SizedIngredient.create(content, modifier.apply(1).intValue());
     }
+
+    @Override
+    public Ingredient deepCopyInner(Ingredient content) {
+        return SizedIngredient.deepCopy(content);
+    }
 }