+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # ChangeLog
 
-## v1.0.28.a
-* Fixes the pattern lock won't release.
+## v1.0.29
+* Fixed Mek EMI support